Refactor hustle view renderers
